apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pymupdf==1.24.10
chromadb==0.5.5
openai==1.44.1
httpx==0.27.2
sentence-transformers==3.0.1
python-dotenv==1.0.1
sentence-transformers==3.0.1